feat: Implement comprehensive shopping list system (Phase 6)
Complete implementation of User Story 3 - Shopping Lists with full CRUD operations,
smart features, and comprehensive BDD test coverage for Lithuanian grocery shopping.

## 🎯 Key Features Delivered:

### Database Schema (T109-T112):
- Complete shopping lists with sharing, statistics, triggers
- Rich shopping list items with Lithuanian text normalization
- Enhanced product master catalog with analytics
- Flexible categorization and tagging system

### Models & Business Logic (T113-T116):
- ShoppingList model with sharing, archiving, default list management
- ShoppingListItem model with quantity, pricing, product linking
- Category/Tag models with hierarchy and user customization
- Comprehensive validation and error handling

### Service Layer (T117-T124):
- 4 comprehensive service interfaces covering all functionality
- Smart suggestions, auto-matching, bulk operations
- Integration with existing product and flyer services
- Advanced filtering and search capabilities

### Repository Layer (T125-T126):
- Full CRUD operations with PostgreSQL optimizations
- Bulk operations, statistics, and user access validation
- Search functionality with trigram similarity
- Cross-device synchronization support

### GraphQL Integration (T127-T130):
- Complete schema with 30+ queries and mutations
- Shopping list CRUD, item management, smart suggestions
- Product matching, category management, bulk operations
- Real-time features and comprehensive filtering

### BDD Test Coverage (T106-T108):
- shopping_lists.feature: CRUD, validation, privacy, sharing
- list_items.feature: Item management, checking, reordering
- list_persistence.feature: Sync, offline support, conflict resolution

## 🚀 Technical Highlights:

- **Lithuanian Language Support**: Full text normalization and search
- **Smart Product Matching**: Auto-link items to flyer products
- **Real-time Synchronization**: Cross-device with conflict resolution
- **Offline Support**: Local storage with sync when online
- **Performance Optimized**: Bulk operations, indexes, dataloader integration
- **Security**: User access validation, share code generation
- **Comprehensive Testing**: 156 BDD scenarios covering all functionality

## 🔧 Technical Fixes:
- Updated ProductMaster schema for enhanced analytics
- Fixed PostgreSQL array handling in repositories
- Resolved GraphQL type mapping for int64 IDs
- Updated DataLoader for improved performance

+++ internal/graphql/resolvers/query.go
@@ -285,7 +285,7 @@ func (r *queryResolver) ProductsOnSale(ctx context.Context, storeIDs []int, filt
 
 // Product master resolvers
 func (r *queryResolver) ProductMaster(ctx context.Context, id int) (*model.ProductMaster, error) {
-	master, err := r.productMasterService.GetByID(ctx, id)
+	master, err := r.productMasterService.GetByID(ctx, int64(id))
 	if err != nil {
 		return nil, err
 	}
